fastapi_project/app/requirements.txt

    fastapi
    uvicorn
    phidata
    google-search-results
    google-generativeai
    openai
    jinja2
    huggingface_hub